requirements.txt
alembic==1.15.2
aniso8601==10.0.1
blinker==1.9.0
click==8.2.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
sentence-transformers==4.1.0 
pdfplumber==0.11.6 
langchain==0.3.25 
openai==1.79.0